fix(orchestration): move ConsoleLogEntry type to shared (#108)
ConsoleLogEntry, ConsoleLogLevel, normalizeConsoleLogEntry, ConsoleLogService,
and defaultConsoleLogService are moved to src/shared/console-log-types.ts.
features/chat/services/console-log.ts re-exports from shared for backward
compatibility. orchestration/agent-loop.ts now imports only from @/shared,
resolving the architecture dependency violation.
